+++ CHANGELOG.md
@@ -6,6 +6,7 @@ WIP
 
 - allow OID4VP for authenticated request
 - rework SD request UI for OID4VC
+- allow custom credential types for OID4VP
 - small improvements